staging. Forward model works.
