--- mbircone/src/denoise3D.h
@@ -1,8 +1,0 @@
-#include "MBIRModularUtilities3D.h"
-#include "icd3d.h"
-#include "icddenoise.h"
-
-
-
-void MBIR3DDenoise(struct Image *img, struct Image *img_noisy, struct ReconParams *reconParams);
-
