+++ mbircone/src/icd3d.h
@@ -94,5 +94,15 @@ void NHICD_checkPartialZiplinesHot(struct ReconAux *reconAux, long int j_x, long
 
 void updateNHICDStats(struct ReconAux *reconAux, long int j_x, long int j_y, struct Image *img, struct ReconParams *reconParams);
 
+// Denoising funcionalities
+void ICDStep3DDenoise(struct Image *img, struct Image *err_image, struct ICDInfo3DCone *icdInfo, struct ReconParams *reconParams);
+
+void computeTheta1Theta2ForwardTermDenoise(struct Image *err_image, struct ICDInfo3DCone *icdInfo, struct ReconParams *reconParams);
+
+float MAPCost3DDenoise(struct Image *image, struct Image *err_image, struct ReconParams *reconParams);
+
+float MAPCostForwardDenoise(struct Image *err_image, struct ReconParams *reconParams);
+
+void disp_iterationInfo_denoise(int itNumber, int MaxIterations, float cost, float relUpdate, float stopThresholdChange, float weightScaler_value, float ticToc_iteration);
 #endif
